Fixed mouse event system for scrollable viewport

+++ project/src/de/topobyte/livecg/core/geometry/geom/Rectangle.java
@@ -38,7 +38,7 @@ public class Rectangle
 		return x1;
 	}
 
-	public void setX1(int x1)
+	public void setX1(double x1)
 	{
 		this.x1 = x1;
 	}
@@ -48,7 +48,7 @@ public class Rectangle
 		return y1;
 	}
 
-	public void setY1(int y1)
+	public void setY1(double y1)
 	{
 		this.y1 = y1;
 	}
@@ -58,7 +58,7 @@ public class Rectangle
 		return x2;
 	}
 
-	public void setX2(int x2)
+	public void setX2(double x2)
 	{
 		this.x2 = x2;
 	}
@@ -68,7 +68,7 @@ public class Rectangle
 		return y2;
 	}
 
-	public void setY2(int y2)
+	public void setY2(double y2)
 	{
 		this.y2 = y2;
 	}
